Add mapper from input model to remote model

+++ features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/common/domain/model/params/add/ProductAddParam.kt
@@ -12,7 +12,7 @@ data class ProductAddParam (
         var productName: String? = null,
         @SerializedName("price")
         @Expose
-        var price: Int? = null,
+        var price: Float? = null,
         @SerializedName("priceCurrency")
         @Expose
         var priceCurrency: String? = null,
